Fix #23: rename silly package name to tofu
